+++ src/router.ts
@@ -1,4 +1,4 @@
-import { createRouter, useParams, type MakeRouteMatch, type Register, type RouteIds } from "@tanstack/react-router";
+import { createRouter } from "@tanstack/react-router";
 
 import ErrorBoundary from "./components/app/templates/error-boundary";
 import PendingBoundary from "./components/app/templates/pending-boundary";
@@ -18,29 +18,3 @@ declare module "@tanstack/react-router" {
 		router: ReturnType<typeof getRouter>;
 	}
 }
-
-export function createRouteSelector<TFrom extends RouteIds<Register["router"]["routeTree"]>, T>(options: { from: TFrom; selector: (route: MakeRouteMatch<Register["router"]["routeTree"], TFrom>) => T }) {
-	return (router: Register["router"]) => {
-		const match = router.state.matches.find((m) => m.routeId === options.from);
-
-		if (!match) {
-			throw new Error(`Tried to access state from unmatched route: ${options.from}`);
-		}
-		return options.selector(match as unknown as MakeRouteMatch<Register["router"]["routeTree"], TFrom>);
-	};
-}
-
-export const selectActiveSongId = createRouteSelector({
-	from: "/_/edit/$sid/$bid/_",
-	selector: ({ params }) => params.sid,
-});
-export const selectActiveBeatmapId = createRouteSelector({
-	from: "/_/edit/$sid/$bid/_",
-	selector: ({ params }) => params.bid,
-});
-export const selectActiveView = createRouteSelector({
-	from: "/_/edit/$sid/$bid/_",
-	selector: ({ context }) => context.view,
-});
-
-useParams;
